requirements.txt
kaggle
kagglehub
polars 
deltalake 
pyarrow 
sqlalchemy 
psycopg2-binary 
python-dotenv 
dagster 
dagit 
dbt-core 
dbt-postgres 
great-expectations 
pytest
pyarrow
dbt
supabase